file restructuring
final tests left
