+++ tests/test_management_fix_no_names.py
@@ -2,7 +2,7 @@ import json
 
 import pdfplumber
 
-from concall_parser.management_fix_no_names import handle_only_management_case
+from concall_parser.management_only import handle_only_management_case
 
 
 def test_handle_only_management_case(filepath: str) -> None:
